Moved ColorPickerActivity to settings package
